+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 ## Unreleased
+* Added liking to videos (+ like count)
 * Fixed bug where first frame of video would be black making thumbnails completly black.
 * Changed default camera to front camera
 * Fixed bug where if you tried going to the front camera and record the app would crash